Removed solenoid test subsystem and added basic methods and structure to launcher subsystem
I deleted the solenoid subsystem for the sake of easy debugging. I didn't want to have to go through and change all the references to variables in the robot map class after changing a large portion of their names. I also updated the Launcher subsystem class which will eventually control the two cylinders that assist the ball in moving through the robot as long as the two flywheel motors on the launching mechanism. Note: there are some runtime errors because of mismatched CAN IDs and some components not being plugged in. I'm either gonna just try every id or research how the CAN system works. WILL FIX