Added rake subsystem and copilot controller support
I added the rake subsystem! Yay! It does stuff. The rake and ball launcher are now controlled by the copilot. D-Pad down lowers the rake and engages the rake motors, D-Pad up does the exact opposite. The "A" button stages the ball (Hold), The right trigger launches the ball unless the motors are off in which case it will engage them. (if this occurs press trigger again to launch ball). If you want to disengages the motors, press the b button. That's basically all I added. Joel, call me on discord if you need help.